0424
favlist update

+++ node_server/src/models/favouriteModel.js
@@ -3,7 +3,7 @@ module.exports = (sequelize, Sequelize) => {
     'favourites',
     {
       favouriteID: {
-        type: Sequelize.INTEGER(11),
+        type: Sequelize.INTEGER(10),
         primaryKey: true,
       },
       customerID: {
